+++ stubs/build.rs
@@ -1,6 +1,8 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    std::env::set_var("PROTOC", protoc_bundled::PROTOC);
-    std::env::set_var("PROTOC_INCLUDE", protoc_bundled::PROTOC_INCLUDE);
+    unsafe {
+        std::env::set_var("PROTOC", protoc_bundled::PROTOC);
+        std::env::set_var("PROTOC_INCLUDE", protoc_bundled::PROTOC_INCLUDE);
+    }
 
     println!("cargo:rerun-if-changed=../protos/dcs");
 
@@ -65,6 +67,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         )
         .build_server(cfg!(feature = "server"))
         .build_client(cfg!(feature = "client"))
-        .compile(&["../protos/dcs/dcs.proto"], &["../protos"])?;
+        .compile_protos(&["../protos/dcs/dcs.proto"], &["../protos"])?;
     Ok(())
 }